@@ -11,25 +11,22 @@
     "e2e": "ng e2e",
     "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && ng serve -o"
   },
-  "engines": {
-    "node": "6.11.1",
-    "npm": "3.10.9"
-  },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.1",
+    "@angular/animations": "14.2.8",
     "@angular/cdk": "^13.3.1",
-    "@angular/common": "^13.3.1",
-    "@angular/compiler": "^13.3.1",
-    "@angular/core": "^13.3.1",
-    "@angular/elements": "^13.3.1",
-    "@angular/forms": "^13.3.1",
+    "@angular/common": "^14.2.8",
+    "@angular/compiler": "^14.2.8",
+    "@angular/core": "^14.2.8",
+    "@angular/elements": "^14.2.8",
+    "@angular/forms": "^14.2.8",
     "@angular/material": "13.3.1",
-    "@angular/platform-browser": "^13.3.1",
-    "@angular/platform-browser-dynamic": "^13.3.1",
-    "@angular/platform-server": "13.3.1",
-    "@angular/router": "^13.3.1",
-    "ajv": "8.11.0",
+    "@angular/platform-browser": "^14.2.8",
+    "@angular/platform-browser-dynamic": "^14.2.8",
+    "@angular/platform-server": "14.2.8",
+    "@angular/router": "^14.2.8",
+    "@aspnet/signalr": "^1.0.27",
+    "ajv": "^8.11.0",
     "arrive": "2.4.1",
     "bootstrap": "4.6.1",
     "bootstrap-material-design": "4.1.3",
@@ -42,7 +39,8 @@
     "googleapis": "99.0.0",
     "hammerjs": "2.0.8",
     "jquery": "3.6.0",
-    "moment": "2.29.1",
+    "moment": "^2.29.4",
+    "ngx-toastr": "^15.2.1",
     "perfect-scrollbar": "1.5.5",
     "popper.js": "1.16.1",
     "rxjs": "^7.5.5",
@@ -51,10 +49,10 @@
     "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.1",
-    "@angular/cli": "^13.3.1",
-    "@angular/compiler-cli": "^13.3.1",
-    "@angular/language-service": "13.3.1",
+    "@angular-devkit/build-angular": "^14.2.7",
+    "@angular/cli": "^14.2.7",
+    "@angular/compiler-cli": "^14.2.8",
+    "@angular/language-service": "14.2.8",
     "@types/bootstrap": "4.5.0",
     "@types/chartist": "0.11.1",
     "@types/google.maps": "3.47.4",
@@ -62,7 +60,7 @@
     "@types/jasminewd2": "~2.0.10",
     "@types/jquery": "3.5.6",
     "@types/node": "16.3.2",
-    "codelyzer": "6.0.2",
+    "codelyzer": "^6.0.2",
     "jasmine-core": "~4.0.1",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "^6.3.17",
@@ -71,10 +69,9 @@
     "karma-coverage-istanbul-reporter": "~3.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "sass": "1.32.13",
     "protractor": "7.0.0",
+    "sass": "1.32.13",
     "ts-node": "~10.7.0",
-    "tslint": "6.1.3",
-    "typescript": "~4.4.2"
+    "typescript": "~4.8.4"
   }
 }
