@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { Country } from 'app/models/country';
 import { State } from 'app/models/state';
 import { HomeService } from 'app/services/home.service';
@@ -13,11 +13,11 @@ export class HomenComponent implements OnInit {
   countries: Country[];
   currentCountry: Country = new Country;
   currentState: State = new State;
-  contactForm:FormGroup;
+  contactForm:UntypedFormGroup;
 
   constructor(
     private homeService: HomeService,
-    private formBuilder:FormBuilder
+    private formBuilder:UntypedFormBuilder
     ) { }
 
   ngOnInit(): void {
